+++ src/Giraffe/TokenRouter.fs
@@ -4,8 +4,8 @@ open System.Collections.Generic
 open System.Text
 open Microsoft.FSharp.Reflection
 open Printf
-open HttpHandlers
-open TokenParsers
+open Giraffe.HttpHandlers
+open Giraffe.TokenParsers
 
 // Implemenation of (router) Trie Node
 // --------------------------------------