+++ tests/Giraffe.Tests/AuthHandlerTests.fs
@@ -65,7 +65,7 @@ module TestApp =
     let private mustBeAdmin = requiresRole "admin" accessDenied
     let private mustBeOperatorOrAdmin = requiresRoleOf ["admin"; "operator"] accessDenied
 
-    let private isJohn (user: ClaimsPrincipal) = user.HasClaim (ClaimTypes.Name, "John")
+    let private isJohn (user : ClaimsPrincipal) = user.HasClaim (ClaimTypes.Name, "John")
     let private mustBeJohn = requiresAuthPolicy isJohn accessDenied
 
     let app =
@@ -185,13 +185,15 @@ module TestData =
 
     type AuthArb =
         static member Values =
-            [anonymousGen
-             nonJohnNoRoleGen
-             nonJohnAdminGen
-             nonJohnOperatorGen
-             johnNoRoleGen
-             johnAdminGen
-             johnOperatorGen]
+            [
+                anonymousGen
+                nonJohnNoRoleGen
+                nonJohnAdminGen
+                nonJohnOperatorGen
+                johnNoRoleGen
+                johnAdminGen
+                johnOperatorGen
+            ]
             |> Gen.oneof
             |> Arb.fromGen