+++ tests/Giraffe.Tests/HttpContextExtensionsTests.fs
@@ -15,6 +15,8 @@ open Microsoft.AspNetCore.TestHost
 open Xunit
 open NSubstitute
 open Newtonsoft.Json
+open Giraffe
+open Giraffe.Serialization
 open Giraffe.GiraffeViewEngine
 
 let assertFailf format args =
@@ -25,6 +27,20 @@ let getBody (ctx : HttpContext) =
     ctx.Response.Body.Position <- 0L
     use reader = new StreamReader(ctx.Response.Body, Encoding.UTF8)
     reader.ReadToEnd()
+
+let mockJson (ctx : HttpContext) (settings : JsonSerializerSettings option) =
+    let jsonSettings =
+        defaultArg settings NewtonsoftJsonSerializer.DefaultSettings
+    ctx.RequestServices
+       .GetService(typeof<IJsonSerializer>)
+       .Returns(NewtonsoftJsonSerializer(jsonSettings))
+    |> ignore
+
+let mockXml (ctx : HttpContext) =
+    ctx.RequestServices
+       .GetService(typeof<IXmlSerializer>)
+       .Returns(DefaultXmlSerializer(DefaultXmlSerializer.DefaultSettings))
+    |> ignore
 
 [<CLIMutable>]
 type ModelWithOption =
@@ -53,6 +69,7 @@ type Customer =
 [<Fact>]
 let ``BindJsonAsync test`` () =
     let ctx = Substitute.For<HttpContext>()
+    mockJson ctx None
 
     let jsonHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
@@ -92,6 +109,7 @@ let ``BindJsonAsync test`` () =
 [<Fact>]
 let ``BindXmlAsync test`` () =
     let ctx = Substitute.For<HttpContext>()
+    mockXml ctx
 
     let xmlHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
@@ -182,7 +200,6 @@ let ``BindQueryString test`` () =
 
     let queryStr = "?Name=John%20Doe&IsVip=true&BirthDate=1990-04-20&Balance=150000.5&LoyaltyPoints=137"
     let query = Microsoft.AspNetCore.WebUtilities.QueryHelpers.ParseQuery queryStr
-    let asdf = QueryCollection(query) :> IQueryCollection
     ctx.Request.Query.ReturnsForAnyArgs(QueryCollection(query) :> IQueryCollection) |> ignore
     ctx.Request.Method.ReturnsForAnyArgs "GET" |> ignore
     ctx.Request.Path.ReturnsForAnyArgs (PathString("/query")) |> ignore
@@ -257,6 +274,7 @@ let ``BindQueryString with option property test`` () =
 [<Fact>]
 let ``BindModelAsync with JSON content returns correct result`` () =
     let ctx = Substitute.For<HttpContext>()
+    mockJson ctx None
 
     let autoHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
@@ -298,11 +316,12 @@ let ``BindModelAsync with JSON content returns correct result`` () =
 [<Fact>]
 let ``BindModelAsync with JSON content that uses custom serialization settings returns correct result`` () =
     let ctx = Substitute.For<HttpContext>()
+    mockJson ctx (Some (JsonSerializerSettings()))
 
     let autoHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
             task {
-                let! model = ctx.BindModelAsync<Customer>(JsonSerializerSettings())
+                let! model = ctx.BindModelAsync<Customer>()
                 return! text (model.ToString()) next ctx
             }
 
@@ -339,6 +358,7 @@ let ``BindModelAsync with JSON content that uses custom serialization settings r
 [<Fact>]
 let ``BindModelAsync with XML content returns correct result`` () =
     let ctx = Substitute.For<HttpContext>()
+    mockXml ctx
 
     let autoHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
@@ -466,6 +486,7 @@ let ``BindModelAsync with culture aware form content returns correct result`` ()
 [<Fact>]
 let ``BindModelAsync with JSON content and a specific charset returns correct result`` () =
     let ctx = Substitute.For<HttpContext>()
+    mockJson ctx None
 
     let autoHandler =
         fun (next : HttpFunc) (ctx : HttpContext) ->
@@ -635,7 +656,7 @@ let ``RenderHtmlAsync should add html to the context`` () =
                 html [] [
                     head [] []
                     body [] [
-                        h1 [] [EncodedText "Hello world"]
+                        h1 [] [ EncodedText "Hello world" ]
                     ]
                 ]
             ctx.RenderHtmlAsync(htmlDoc)