+++ RELEASE_NOTES.md
@@ -1,6 +1,26 @@
 Release Notes
 =============
 
+## 0.1.0-beta-800
+
+#### Breaking changes
+
+- Removed serialization methods from `Giraffe.Common`
+    - Removed `serializeJson`, `deserializeJson<'T>`, `deserializeJsonFromStream<'T>`, `defaultJsonSerializerSettings`, `defaultSerializeJson`, `defaultDeserializeJson<'T>`, `serializeXml` and `deserializeXml<'T>`
+- Removed the overloards for `BindJsonAsync<'T>`, `BindModelAsync<'T>` and `WriteJsonAsync` which accepted an object of type `JsonSerializerSettings`
+- Removed the `customJson` http handler function
+- In order for Giraffe to be fully functional one has to register the default Giraffe dependencies by calling `services.AddGiraffe()` from your `RegisterServices` method in the startup code.
+
+#### New features
+
+- Added `GetJsonSerializer()` and `GetXmlSerializer()` to  `Giraffe.HttpContextExtensions`
+- Added a new `INegotiationConfig` interface which can be used to configure the behaviour of the `negotiate` http handler
+- Added a new `IJsonSerializer` interface which can be used to configure JSON serialization in Giraffe
+- Added the `NewtonsoftJsonSerializer` class which implements `IJsonSerializer` and gets auto registered when adding the Giraffe dependencies via `services.AddGiraffe()` in your startup code.
+- Added a new `IXmlSerializer` interface which can be used to configure XML serialization in Giraffe
+- Added the `DefaultXmlSerializer` class which implements `IXmlSerializer` and gets auto registered when adding the Giraffe dependencies via `services.AddGiraffe()` in your startup code.
+
+
 ## 0.1.0-beta-700
 
 #### Breaking changes